Recent changes

+++ Range Minimum Query(Sparse Table).cpp
@@ -21,44 +21,44 @@ template<typename T>T absll(T X)
         return X;
 }
 
-int logT[MAXN];
-int rmq[33][MAXN];
-int PTwo[33];
-int A[MAXN];
+LL logT[MAXN];
+LL rmq[33][MAXN];
+LL PTwo[33];
+LL A[MAXN];
 
-void RMQSparseTable(int N)
+void RMQSparseTable(LL N)
 {
 	logT[1]=0;
 	PTwo[0]=1;
-	for(int i=1;i<30;i++)
+	for(LL i=1;i<30;i++)
 	{
 		PTwo[i]=PTwo[i-1]*2;
 	}
-	for(int i=2;i<=N;i++)
+	for(LL i=2;i<=N;i++)
 	{
 		logT[i]=logT[i>>1]+1;
 	}
 	
-	for(int i=0;i<N;i++)
+	for(LL i=0;i<N;i++)
 	{
 		rmq[0][i]=i;
 	}
 	
-	for(int k=1;PTwo[k]<N;k++)
+	for(LL k=1;PTwo[k]<N;k++)
 	{
-		for(int i=0;i+PTwo[k]<=N;i++)
+		for(LL i=0;i+PTwo[k]<=N;i++)
 		{
-			int X=rmq[k-1][i];
-			int Y=rmq[k-1][i+PTwo[k-1]];
+			LL X=rmq[k-1][i];
+			LL Y=rmq[k-1][i+PTwo[k-1]];
 			rmq[k][i]=(A[X]<=A[Y]?X:Y);
 		}
 	}
 }
 
-int RMQQuery(int i,int j)
+LL RMQQuery(LL i,LL j)
 {
-	int k=logT[j-i];
-	int x=rmq[k][i];
-	int y=rmq[k][j-PTwo[k]+1];
+	LL k=logT[j-i];
+	LL x=rmq[k][i];
+	LL y=rmq[k][j-PTwo[k]+1];
 	return A[x]<=A[y]?x:y;
 }
